requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
torch>=1.11.0
torchvision>=0.12.0

# Network Analysis
scapy>=2.4.5
pyshark>=0.4.6
netaddr>=0.8.0

# Data Visualization and Dashboard
streamlit>=1.12.0
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data Processing and Utils
joblib>=1.1.0
imbalanced-learn>=0.8.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Monitoring and Logging
psutil>=5.8.0
tqdm>=4.64.0

# Configuration and Environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Development and Testing
pytest>=7.0.0
jupyter>=1.0.0
ipython>=8.0.0

# Optional: Deep Learning acceleration
# tensorflow-gpu>=2.8.0  # Uncomment if using GPU
# torch-audio>=0.11.0    # Uncomment if needed